@@ -1,11 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net462</TargetFramework>
-		<LangVersion>9</LangVersion>
+		<TargetFramework>net7.0</TargetFramework>
+		<LangVersion>11</LangVersion>
 		<AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
 		<OutputPath>bin\$(Configuration)\Mods\mod</OutputPath>
 		<RootNamespace>RPVoiceChat</RootNamespace>
+		<CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
 	</PropertyGroup>
 
 	<ItemGroup>
@@ -33,9 +34,13 @@
 			<HintPath>$(VINTAGE_STORY)/Mods/VSEssentials.dll</HintPath>
 			<Private>False</Private>
 		</Reference>
-		<Reference Include="OpenTK">
-			<HintPath>$(VINTAGE_STORY)/Lib/OpenTK.dll</HintPath>
-			<Private>False</Private>
+		<Reference Include="OpenAL">
+			<HintPath>Lib/OpenTK.Audio.OpenAL.dll</HintPath>
+			<Private>True</Private>
+		</Reference>
+		<Reference Include="OpenTK.Mathematics">
+			<HintPath>Lib/OpenTK.Mathematics.dll</HintPath>
+			<Private>True</Private>
 		</Reference>
 		<Reference Include="Newtonsoft.Json">
 			<HintPath>$(VINTAGE_STORY)/Lib/Newtonsoft.Json.dll</HintPath>
@@ -74,7 +79,6 @@
 
 	<ItemGroup>
 		<PackageReference Include="Concentus" Version="1.1.7" />
-		<PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
 		<PackageReference Include="NETStandard.Library" Version="2.0.3" />
 	</ItemGroup>
 </Project>
