@@ -5,6 +5,6 @@ namespace RPVoiceChat.Gui
     public interface IExtendedGuiElement
     {
         public void Init(string elementKey, ElementBounds bounds, GuiComposer composer);
-        public void OnAdd(GuiComposer composer);
+        public void OnAdd(GuiComposer composer) { }
     }
 }
