@@ -1,5 +1,8 @@
+using HarmonyLib;
 using System;
+using System.Reflection;
 using Vintagestory.API.Server;
+using Vintagestory.Server;
 
 namespace RPVoiceChat.Networking
 {
@@ -8,11 +11,25 @@ namespace RPVoiceChat.Networking
         public event Action<AudioPacket> AudioPacketReceived;
         private ICoreServerAPI api;
         private IServerNetworkChannel channel;
+        private ServerSystemNetworkProcess networkProcess;
 
         public NativeNetworkServer(ICoreServerAPI sapi) : base(sapi)
         {
             api = sapi;
-            channel = sapi.Network.GetChannel(ChannelName).SetMessageHandler<AudioPacket>(ReceivedAudioPacketFromClient);
+            channel = api.Network.GetChannel(ChannelName).SetMessageHandler<AudioPacket>(ReceivedAudioPacketFromClient);
+            if (api.Server.IsDedicated == false)
+                api.Network.RegisterChannel(SPChannelName)
+                    .RegisterMessageType<AudioPacket>()
+                    .SetMessageHandler<AudioPacket>(ReceivedAudioPacketFromClient);
+
+            NetworkAPIPatch.OnHandleCustomPacket += ShouldProcessInBackground;
+        }
+
+        public void Launch()
+        {
+            networkProcess = new ServerSystemNetworkProcess(api);
+            networkProcess.OnProcessInBackground += ProcessInBackground;
+            networkProcess.Launch();
         }
 
         public ConnectionInfo GetConnectionInfo()
@@ -28,9 +45,32 @@ namespace RPVoiceChat.Networking
             return true;
         }
 
+        private bool ProcessInBackground(int channelId, Packet_CustomPacket customPacket, IServerPlayer sender)
+        {
+            if (ShouldProcessInBackground(channelId) == false) return false;
+
+            ((NetworkChannel)channel).OnPacket(customPacket, sender);
+            return true;
+        }
+
+        private static FieldInfo channelIdField = AccessTools.Field(typeof(NetworkChannel), "channelId");
+
+        private bool ShouldProcessInBackground(int channelId)
+        {
+            if (channel is not NetworkChannel) return false;
+            var expectedChannelId = (int)channelIdField.GetValue(channel);
+            return channelId == expectedChannelId;
+        }
+
         private void ReceivedAudioPacketFromClient(IServerPlayer player, AudioPacket packet)
         {
             AudioPacketReceived?.Invoke(packet);
+        }
+
+        public void Dispose()
+        {
+            NetworkAPIPatch.OnHandleCustomPacket -= ShouldProcessInBackground;
+            networkProcess?.Dispose();
         }
     }
 }
