@@ -7,15 +7,14 @@ namespace RPVoiceChat.Utils
     {
         public static int ChannelsPerFormat(ALFormat format)
         {
-            switch (format)
-            {
-                case ALFormat.Mono16:
-                    return 1;
-                case ALFormat.MultiQuad16Ext:
-                    return 4;
-                default:
-                    throw new NotSupportedException($"Format {format} is not supported for capture");
-            }
+            int channels = format switch {
+                ALFormat.Mono16 => 1,
+                ALFormat.Stereo16 => 2,
+                ALFormat.MultiQuad16Ext => 4,
+                _ => throw new NotSupportedException($"Format {format} is not supported for capture")
+            };
+
+            return channels;
         }
 
         public static byte[] ShortsToBytes(short[] audio, int offset, int length)
